rename carousel files accordingly, delete slider folder and files
