+++ client/components/Explore/Carousel.jsx
@@ -1,60 +1,33 @@
 import * as React from 'react';
-import { useCarousel } from './useCarousel';
+import { CarouselContainer, CarouselChild } from './CarouselContainer';
 
-function makeIndices(start, delta, num) {
-    const indices = [];
-
-    while (indices.length < num) {
-        indices.push(start);
-        start += delta;
-    }
-
-    return indices;
-}
-
-const CarouselContainer = ({
-    children,
-    slidesPresented = 1,
-    interval = 5000,
-}) => {
-    const slides = React.Children.toArray(children);
-    const length = slides.length;
-    const numActive = Math.min(length, slidesPresented);
-    const [active, setActive, handlers, style] = useCarousel(length, interval, { slidesPresented: numActive });
-    const beforeIndices = makeIndices(slides.length - 1, -1, numActive);
-    const afterIndices = makeIndices(0, +1, numActive);
-
-    return (
-        length > 0 && (
-            <div className="carousel">
-                <ol className="carousel-indicators">
-                    {slides.map((_, index) => (
-                        <li
-                            onClick={() => setActive(index)}
-                            key={index}
-                            className={`${active === index ? 'active' : ''} carousel-indicator`}
-                        />
-                    ))}
-                </ol>
-                <div className="carousel-content" {...handlers} style={style}>
-                    {beforeIndices.map(i => (
-                        <CarouselChild key={i}>{slides[i]}</CarouselChild>
-                    ))}
-                    {slides.map((slide, index) => (
-                        <CarouselChild key={index}>{slide}</CarouselChild>
-                    ))}
-                    {afterIndices.map(i => (
-                        <CarouselChild key={i}>{slides[i]}</CarouselChild>
-                    ))}
-                </div>
-            </div>
-        )
-    );
-};
-
-
-const CarouselChild = ({ children }) => (
-    <div className="carousel-item">{children}</div>
+const Carousel = () => (
+    <CarouselContainer interval={5000}>
+        <CarouselChild>
+            <div
+                className="slide"
+                style={{ backgroundImage: `url(https://gatherplatform.s3.us-east-2.amazonaws.com/photo-1.png)` }}
+            />
+        </CarouselChild>
+        <CarouselChild>
+            <div
+                className="slide"
+                style={{ backgroundImage: `url(https://gatherplatform.s3.us-east-2.amazonaws.com/photo-2.jpg)` }}
+            />
+        </CarouselChild>
+        <CarouselChild>
+            <div
+                className="slide"
+                style={{ backgroundImage: `url(https://gatherplatform.s3.us-east-2.amazonaws.com/photo-3.png)` }}
+            />
+        </CarouselChild>
+        <CarouselChild>
+            <div
+                className="slide"
+                style={{ backgroundImage: `url(https://gatherplatform.s3.us-east-2.amazonaws.com/photo-4.png)` }}
+            />
+        </CarouselChild>
+    </CarouselContainer>
 );
 
-export { CarouselChild, CarouselContainer };
+export default Carousel;
